+++ tests/Rules/Data/CorrectDeferrableProviderIndirect.php
@@ -1,0 +1,10 @@
+<?php
+declare(strict_types=1);
+
+namespace Rules\Data;
+
+use Illuminate\Contracts\Support\DeferrableProvider;
+
+class CorrectDeferrableProviderIndirect extends CorrectDeferrableProvider implements DeferrableProvider
+{
+}